+++ src/main/java/net/minecraftforge/common/model/animation/IAnimationStateMachine.java
@@ -34,7 +34,7 @@ public interface IAnimationStateMachine
      * Event iterable will contain all events that happened from the last invocation of this method, from most to least recent.
      * Event offset is relative to the previous event, and for the first event it's relative to the current time. 
      */
-    Pair<IModelState, Iterable<Event>> apply(float time);
+    Pair<IModelState, Iterable<Event>> apply(double time);
 
     /**
      * Transition to a new state.
